+++ mne/io/meas_info.py
@@ -11,6 +11,7 @@ from scipy import linalg
 from ..externals.six import BytesIO, string_types
 from datetime import datetime as dt
 
+from .pick import channel_type
 from .constants import FIFF
 from .open import fiff_open
 from .tree import dir_tree_find, copy_tree
@@ -22,6 +23,7 @@ from .write import (start_file, end_file, start_block, end_block,
                     write_coord_trans, write_ch_info, write_name_list,
                     write_julian)
 from ..utils import logger, verbose
+from ..fixes import Counter
 
 _kind_dict = dict(
     eeg=(FIFF.FIFFV_EEG_CH, FIFF.FIFFV_COIL_NONE, FIFF.FIFF_UNIT_V),
@@ -68,12 +70,18 @@ class Info(dict):
                 entr = dt.fromtimestamp(v[0]).strftime('%Y-%m-%d %H:%M:%S')
             else:
                 this_len = (len(v) if hasattr(v, '__len__') else
-                           ('%s' % v if v is not None else None))
+                            ('%s' % v if v is not None else None))
                 entr = (('%d items' % this_len) if isinstance(this_len, int)
                         else ('%s' % this_len if this_len else ''))
             if entr:
                 non_empty += 1
                 entr = ' | ' + entr
+            if k == 'chs':
+                ch_types = [channel_type(self, idx) for idx in range(len(v))]
+                ch_counts = Counter(ch_types)
+                entr += " (%s)" % ', '.join("%s: %d" % (ch_type.upper(), count)
+                                            for ch_type, count
+                                            in ch_counts.items())
             strs.append('%s : %s%s' % (k, str(type(v))[7:-2], entr))
         strs_non_empty = sorted(s for s in strs if '|' in s)
         strs_empty = sorted(s for s in strs if '|' not in s)
@@ -282,10 +290,10 @@ def read_meas_info(fid, tree, verbose=None):
             tag = read_tag(fid, pos)
             cand = tag.data
             if cand['from'] == FIFF.FIFFV_COORD_DEVICE and \
-                                cand['to'] == FIFF.FIFFV_COORD_HEAD:
+                    cand['to'] == FIFF.FIFFV_COORD_HEAD:
                 dev_head_t = cand
             elif cand['from'] == FIFF.FIFFV_MNE_COORD_CTF_HEAD and \
-                                cand['to'] == FIFF.FIFFV_COORD_HEAD:
+                    cand['to'] == FIFF.FIFFV_COORD_HEAD:
                 ctf_head_t = cand
         elif kind == FIFF.FIFF_EXPERIMENTER:
             tag = read_tag(fid, pos)
@@ -327,8 +335,8 @@ def read_meas_info(fid, tree, verbose=None):
                     tag = read_tag(fid, pos)
                     cand = tag.data
                     if (cand['from'] == FIFF.FIFFV_COORD_DEVICE and
-                        cand['to'] == FIFF.FIFFV_COORD_HEAD and
-                        dev_head_t is None):
+                            cand['to'] == FIFF.FIFFV_COORD_HEAD and
+                            dev_head_t is None):
                         dev_head_t = cand
                     elif (cand['from'] == FIFF.FIFFV_MNE_COORD_CTF_HEAD and
                           cand['to'] == FIFF.FIFFV_COORD_HEAD and
@@ -659,7 +667,7 @@ def _is_equal_dict(dicts):
     for d in tests:
         k0, v0 = d[0]
         is_equal.append(all([np.all(k == k0) and
-                        np.all(v == v0)  for k, v in d]))
+                        np.all(v == v0) for k, v in d]))
     return all(is_equal)
 
 
@@ -695,7 +703,7 @@ def _merge_dict_values(dicts, key, verbose=None):
             return values[int(idx)]
         elif len(idx) > 1:
             lists = (d[key] for d in dicts if isinstance(d[key], list))
-            return  _flatten(lists)
+            return _flatten(lists)
     # dict
     elif _check_isinstance(values, dict, all):
         is_qual = _is_equal_dict(values)